Fix bug where newlines are inserted into asset
PHAsset's `requestDataForAssetResource` calls `dataReceivedHandler` multiple times, with a portion of the asset's data each time. The receiver is expected to append this data together to complete the asset.

MultipartDataBuilder was incorrectly inserting a newline character each time `dataReceivedHandler` was being called, causing corruption of the asset's data.

To fix this, we remove the newline writing from `dataReceivedHandler`, and move it down to the `completionHandler`, thus acheiving the desired output, which is a single newline after the asset's entire, complete data.