Switch from CommonCrypto to CryptoSwift
The CommonCrypto module mapping is no longer functioning in xcode 9. The problem is that xcode 9 now requires specificly pointing to the simulator SDK when building to simulator, and won't function with the generic iOS sdk, and cocoapods doesn't support the simulator as a target platform.
Instead of spending tons of time trying to figure out a way around this, I'm switching the md5 function's usage to an alternate, pure-swift library.